# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wheels API",
        "description": "API for wheels app",
        "contact": {
            "email": "adrian.sullca@cirvianum.cat"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/announcements": {
            "get": {
                "tags": [
                    "Announcements"
                ],
                "summary": "Retrieve all active announcements except those created by the authenticated user",
                "operationId": "ce475728e053e113e721a129a915c22b",
                "responses": {
                    "200": {
                        "description": "List of announcements retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/announcements/create": {
            "post": {
                "tags": [
                    "Announcements"
                ],
                "summary": "Create a new announcement",
                "description": "Creates a new announcement with details such as title, description, model, brand, year, and associated photos.",
                "operationId": "d62a861bc49dbb3944847a10705e27a8",
                "requestBody": {
                    "description": "Announcement details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Brand new sedan"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Low mileage, well-maintained car."
                                    },
                                    "model": {
                                        "type": "string",
                                        "example": "Corolla"
                                    },
                                    "brand": {
                                        "type": "string",
                                        "example": "Toyota"
                                    },
                                    "kilometers": {
                                        "type": "number",
                                        "example": 15000
                                    },
                                    "year": {
                                        "type": "integer",
                                        "example": 2022
                                    },
                                    "price": {
                                        "type": "number",
                                        "example": 25000
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "active"
                                    },
                                    "photos": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "vehicleType": {
                                        "type": "string",
                                        "example": "sedan"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Announcement created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Announcement created successfully"
                                        },
                                        "announcement": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "title": {
                                                    "type": "string",
                                                    "example": "Brand new sedan"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Low mileage, well-maintained car."
                                                },
                                                "model": {
                                                    "type": "string",
                                                    "example": "Corolla"
                                                },
                                                "brand": {
                                                    "type": "string",
                                                    "example": "Toyota"
                                                },
                                                "kilometers": {
                                                    "type": "number",
                                                    "example": 15000
                                                },
                                                "year": {
                                                    "type": "integer",
                                                    "example": 2022
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "example": 25000
                                                },
                                                "state": {
                                                    "type": "string",
                                                    "example": "active"
                                                },
                                                "vehicleType": {
                                                    "type": "string",
                                                    "example": "sedan"
                                                },
                                                "photoUrls": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "https://res.cloudinary.com/example/image/upload/photo1.jpg"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Authentication required. Please log in."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}